Add "[" and "]" commands.
"[" Will jump past the matching "]" if the current cell is zero. Otherwise, it does nothing. "]" will jump back to the matching "[" if the current cell is not zero. Otherwise, it does nothing. This enables while loops in BF.

Because flow control means we will be running a BF program for an unknown amount of time, the for loop has been changed to a while loop which exits after reacbing the end of the program.In addition, "program_string" was changed to "program" because it's cleaner looking.