Change get_brace_matches to include right brace keys and left brace values.
Before, we were just storing left brace keys to right brace values. Now we store both ways. This allows us to not preform a linear search when trying to jump backwards in a "]" command. Note that this does mean we store twice as many key-value pairs.

Tests have been updated to reflect this new change.